feat(frontend): Add button function passthrough as prop to SectionHeaders
Function for the button click event of the SectionHeader is provided by the parent component. The parent provdes the function as through the component props and the child calls the function provided in the prop. The same passthrough was added to the RowComponent, thus a user of the RowComponent can provide a function for the button without requireing knowledge about the internal structure of the component.